dashboard-v2/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic[email]
pydantic-settings==2.1.0
kubernetes-asyncio==28.2.1
redis==5.0.1
python-socketio==5.11.0
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
pyyaml==6.0.1
